fix: Resolve HDF5 build failures on modern compilers
Four issues from PR #23, plus pre-existing warnings surfaced by the
now-compiling HDF5 code paths:

Makefile: Replace hardcoded HDF5 path with pkg-config auto-detection.
Uses 'pkg-config --variable=prefix hdf5_hl' with a fallback to hdf5
and finally /usr/local, so any standard installation is found without
editing the Makefile. Add -lhdf5_hl to HDF5LIB; the High-Level Table
API (H5TBmake_table, H5TBappend_records) is in a separate library on
modern HDF5 distributions. HDF5DIR can still be overridden manually:
make USE-HDF5=yes HDF5DIR=/custom/path. USE-HDF5 remains opt-in
(commented out) and CC remains cc.

code/core_simulation.h: Remove *Halo from the struct tail definition.
The pattern '} *Halo;' creates a tentative definition in every
translation unit that includes the header, causing duplicate symbol
errors on strict linkers (Apple ld). Halo is already declared extern
in globals.h and defined once in core_allvars.c.

code/io_save_hdf5.c: Add missing #include "util_parameters.h".
store_run_properties() uses ParameterDefinition, get_parameter_table(),
and get_parameter_table_size() from that header. Also suppress five
pre-existing 'set but not used' warnings on the herr_t/hid_t status
variables in functions that intentionally discard HDF5 return values.

code/io_tree_hdf5.c: Fix format specifier %d -> %lld for hid_t, which
is long long on HDF5 >= 1.10.

Fixes reported in PR #23. CC and USE-HDF5 default behaviour unchanged.

+++ code/io_tree_hdf5.c
@@ -205,8 +205,8 @@ void load_tree_hdf5(int32_t filenr, int32_t treenr) {
     char err_msg[MAX_STRING_LEN + 1];
     snprintf(err_msg, MAX_STRING_LEN,
              "The HDF5 file should still be opened when reading the halos in "
-             "tree %d. Error code: %d",
-             treenr, hdf5_file);
+             "tree %d. Error code: %lld",
+             treenr, (long long)hdf5_file);
     fprintf(stderr, "%s\n", err_msg);
     ABORT(0);
   }
